Set instance-state and instance-type.
This patch adds a couple of pieces of metadata to Machines to expose
some useful information as a label and an annotation.  Some generic
names are being used here to not be so tied to the details of the
BareMetalHost.

The instance-state annotation is set to the provisioning status of the
underlying BareMetalHost. The instance-type label is set to the
hardware profile of the underlying BareMetalHost.